Split up default multiplexer file
