+++ libcaf_io/caf/io/network/doorman_impl.hpp
@@ -18,45 +18,41 @@
 
 #pragma once
 
+#include "caf/io/fwd.hpp"
+#include "caf/io/doorman.hpp"
+
+#include "caf/io/network/acceptor_impl.hpp"
+#include "caf/io/network/native_socket.hpp"
+
+#include "caf/io/network/policy/tcp.hpp"
+
 namespace caf {
-
-// -- templates from the parent namespace necessary for defining aliases -------
-
-template <class> class intrusive_ptr;
-
 namespace io {
-
-// -- variadic templates -------------------------------------------------------
-
-template <class... Sigs>
-class typed_broker;
-
-// -- classes ------------------------------------------------------------------
-
-class scribe;
-class broker;
-class doorman;
-class middleman;
-class basp_broker;
-class receive_policy;
-class abstract_broker;
-class datagram_servant;
-
-// -- aliases ------------------------------------------------------------------
-
-using scribe_ptr = intrusive_ptr<scribe>;
-using doorman_ptr = intrusive_ptr<doorman>;
-using datagram_servant_ptr = intrusive_ptr<datagram_servant>;
-
-// -- nested namespaces --------------------------------------------------------
-
 namespace network {
 
-class multiplexer;
-class default_multiplexer;
-
+/// Default doorman implementation.
+class doorman_impl : public doorman {
+public:
+  doorman_impl(default_multiplexer& mx, native_socket sockfd);
+  
+  bool new_connection() override;
+  
+  void stop_reading() override;
+  
+  void launch() override;
+  
+  std::string addr() const override;
+  
+  uint16_t port() const override;
+  
+  void add_to_loop() override;
+  
+  void remove_from_loop() override;
+  
+protected:
+  acceptor_impl<policy::tcp> acceptor_;
+};
+  
 } // namespace network
-
 } // namespace io
 } // namespace caf
-
