+++ libcaf_io/caf/io/network/event_handler.hpp
@@ -1,0 +1,91 @@
+/******************************************************************************
+ *                       ____    _    _____                                   *
+ *                      / ___|  / \  |  ___|    C++                           *
+ *                     | |     / _ \ | |_       Actor                         *
+ *                     | |___ / ___ \|  _|      Framework                     *
+ *                      \____/_/   \_|_|                                      *
+ *                                                                            *
+ * Copyright 2011-2018 Dominik Charousset                                     *
+ *                                                                            *
+ * Distributed under the terms and conditions of the BSD 3-Clause License or  *
+ * (at your option) under the terms and conditions of the Boost Software      *
+ * License 1.0. See accompanying files LICENSE and LICENSE_ALTERNATIVE.       *
+ *                                                                            *
+ * If you did not receive a copy of the license files, see                    *
+ * http://opensource.org/licenses/BSD-3-Clause and                            *
+ * http://www.boost.org/LICENSE_1_0.txt.                                      *
+ ******************************************************************************/
+
+#pragma once
+
+#include "caf/io/fwd.hpp"
+
+#include "caf/io/network/operation.hpp"
+#include "caf/io/network/native_socket.hpp"
+
+namespace caf {
+namespace io {
+namespace network {
+
+/// A socket I/O event handler.
+class event_handler {
+public:
+  event_handler(default_multiplexer& dm, native_socket sockfd);
+  
+  virtual ~event_handler();
+  
+  /// Returns true once the requested operation is done, i.e.,
+  /// to signalize the multiplexer to remove this handler.
+  /// The handler remains in the event loop as long as it returns false.
+  virtual void handle_event(operation op) = 0;
+  
+  /// Callback to signalize that this handler has been removed
+  /// from the event loop for operations of type `op`.
+  virtual void removed_from_loop(operation op) = 0;
+  
+  /// Returns the native socket handle for this handler.
+  inline native_socket fd() const {
+    return fd_;
+  }
+  
+  /// Returns the `multiplexer` this acceptor belongs to.
+  inline default_multiplexer& backend() {
+    return backend_;
+  }
+  
+  /// Returns the bit field storing the subscribed events.
+  inline int eventbf() const {
+    return eventbf_;
+  }
+  
+  /// Sets the bit field storing the subscribed events.
+  inline void eventbf(int value) {
+    eventbf_ = value;
+  }
+  
+  /// Checks whether `close_read` has been called.
+  inline bool read_channel_closed() const {
+    return read_channel_closed_;
+  }
+  
+  /// Closes the read channel of the underlying socket.
+  void close_read_channel();
+  
+  /// Removes the file descriptor from the event loop of the parent.
+  void passivate();
+  
+protected:
+  /// Adds the file descriptor to the event loop of the parent.
+  void activate();
+  
+  void set_fd_flags();
+  
+  int eventbf_;
+  native_socket fd_;
+  bool read_channel_closed_;
+  default_multiplexer& backend_;
+};
+  
+} // namespace network
+} // namespace io
+} // namespace caf
