+++ libcaf_io/caf/io/network/policy/udp.hpp
@@ -18,45 +18,26 @@
 
 #pragma once
 
+#include "caf/io/network/default_multiplexer.hpp"
+
 namespace caf {
-
-// -- templates from the parent namespace necessary for defining aliases -------
-
-template <class> class intrusive_ptr;
-
 namespace io {
-
-// -- variadic templates -------------------------------------------------------
-
-template <class... Sigs>
-class typed_broker;
-
-// -- classes ------------------------------------------------------------------
-
-class scribe;
-class broker;
-class doorman;
-class middleman;
-class basp_broker;
-class receive_policy;
-class abstract_broker;
-class datagram_servant;
-
-// -- aliases ------------------------------------------------------------------
-
-using scribe_ptr = intrusive_ptr<scribe>;
-using doorman_ptr = intrusive_ptr<doorman>;
-using datagram_servant_ptr = intrusive_ptr<datagram_servant>;
-
-// -- nested namespaces --------------------------------------------------------
-
 namespace network {
+namespace policy {
 
-class multiplexer;
-class default_multiplexer;
+/// Function signature of read_datagram
+using read_datagram_fun = decltype(read_datagram)*;
 
+/// Function signature of write_datagram
+using write_datagram_fun = decltype(write_datagram)*;
+
+/// Policy object for wrapping default UDP operations
+struct udp {
+  static read_datagram_fun read_datagram;
+  static write_datagram_fun write_datagram;
+};
+
+} // namespace policy
 } // namespace network
-
 } // namespace io
 } // namespace caf
-
