+++ libcaf_io/src/default_multiplexer.cpp
@@ -18,15 +18,26 @@
 
 #include "caf/io/network/default_multiplexer.hpp"
 
-#include "caf/actor_system_config.hpp"
 #include "caf/config.hpp"
 #include "caf/defaults.hpp"
+#include "caf/optional.hpp"
+#include "caf/make_counted.hpp"
+#include "caf/actor_system_config.hpp"
+
 #include "caf/io/broker.hpp"
 #include "caf/io/middleman.hpp"
-#include "caf/io/network/interfaces.hpp"
+#include "caf/io/network/stream.hpp"
+#include "caf/io/network/acceptor.hpp"
 #include "caf/io/network/protocol.hpp"
-#include "caf/make_counted.hpp"
-#include "caf/optional.hpp"
+#include "caf/io/network/interfaces.hpp"
+#include "caf/io/network/pipe_reader.hpp"
+#include "caf/io/network/scribe_impl.hpp"
+#include "caf/io/network/doorman_impl.hpp"
+#include "caf/io/network/datagram_handler.hpp"
+#include "caf/io/network/datagram_servant_impl.hpp"
+#include "caf/io/network/policy/tcp.hpp"
+#include "caf/io/network/policy/udp.hpp"
+
 #include "caf/scheduler/abstract_coordinator.hpp"
 
 #ifdef CAF_WINDOWS
@@ -48,11 +59,7 @@
 
 using std::string;
 
-// -- Utiliy functions for converting errno into CAF errors --------------------
-
 namespace {
-
-constexpr size_t receive_buffer_size = std::numeric_limits<uint16_t>::max();
 
 // safe ourselves some typing
 constexpr auto ipv4 = caf::io::network::protocol::ipv4;
@@ -66,11 +73,6 @@ bool cc_zero(int value) {
 // predicate for `ccall` meaning "expected result of f is 1"
 bool cc_one(int value) {
   return value == 1;
-}
-
-// predicate for `ccall` meaning "expected result of f is not -1"
-bool cc_not_minus1(int value) {
-  return value != -1;
 }
 
 // predicate for `ccall` meaning "expected result of f is a valid socket"
@@ -105,180 +107,6 @@ bool cc_valid_socket(caf::io::network::native_socket fd) {
 namespace caf {
 namespace io {
 namespace network {
-
-// -- OS-specific functions for sockets and pipes ------------------------------
-
-#ifndef CAF_WINDOWS
-
-  string last_socket_error_as_string() {
-    return strerror(errno);
-  }
-
-  expected<void> nonblocking(native_socket fd, bool new_value) {
-    CAF_LOG_TRACE(CAF_ARG(fd) << CAF_ARG(new_value));
-    // read flags for fd
-    CALL_CFUN(rf, cc_not_minus1, "fcntl", fcntl(fd, F_GETFL, 0));
-    // calculate and set new flags
-    auto wf = new_value ? (rf | O_NONBLOCK) : (rf & (~(O_NONBLOCK)));
-    CALL_CFUN(set_res, cc_not_minus1, "fcntl", fcntl(fd, F_SETFL, wf));
-    return unit;
-  }
-
-  expected<void> allow_sigpipe(native_socket fd, bool new_value) {
-    if (no_sigpipe_socket_flag != 0) {
-      int value = new_value ? 0 : 1;
-      CALL_CFUN(res, cc_zero, "setsockopt",
-                setsockopt(fd, SOL_SOCKET, no_sigpipe_socket_flag, &value,
-                           static_cast<unsigned>(sizeof(value))));
-    }
-    return unit;
-  }
-
-  expected<void> allow_udp_connreset(native_socket, bool) {
-    // nop; SIO_UDP_CONNRESET only exists on Windows
-    return unit;
-  }
-
-  std::pair<native_socket, native_socket> create_pipe() {
-    int pipefds[2];
-    if (pipe(pipefds) != 0) {
-      perror("pipe");
-      exit(EXIT_FAILURE);
-    }
-    return {pipefds[0], pipefds[1]};
-  }
-
-#else // CAF_WINDOWS
-
-  string last_socket_error_as_string() {
-    LPTSTR errorText = NULL;
-    auto hresult = last_socket_error();
-    FormatMessage( // use system message tables to retrieve error text
-      FORMAT_MESSAGE_FROM_SYSTEM
-      // allocate buffer on local heap for error text
-      | FORMAT_MESSAGE_ALLOCATE_BUFFER
-      // Important! will fail otherwise, since we're not
-      // (and CANNOT) pass insertion parameters
-      | FORMAT_MESSAGE_IGNORE_INSERTS,
-      nullptr, // unused with FORMAT_MESSAGE_FROM_SYSTEM
-      hresult, MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT),
-      (LPTSTR) & errorText, // output
-      0,                    // minimum size for output buffer
-      nullptr);             // arguments - see note
-    std::string result;
-    if (errorText != nullptr) {
-      result = errorText;
-      // release memory allocated by FormatMessage()
-      LocalFree(errorText);
-    }
-    return result;
-  }
-
-  expected<void> nonblocking(native_socket fd, bool new_value) {
-    u_long mode = new_value ? 1 : 0;
-    CALL_CFUN(res, cc_zero, "ioctlsocket", ioctlsocket(fd, FIONBIO, &mode));
-    return unit;
-  }
-
-  expected<void> allow_sigpipe(native_socket, bool) {
-    // nop; SIGPIPE does not exist on Windows
-    return unit;
-  }
-
-  expected<void> allow_udp_connreset(native_socket fd, bool new_value) {
-    DWORD bytes_returned = 0;
-    CALL_CFUN(res, cc_zero, "WSAIoctl",
-              WSAIoctl(fd, SIO_UDP_CONNRESET, &new_value, sizeof(new_value),
-                       NULL, 0, &bytes_returned, NULL, NULL));
-    return unit;
-  }
-
-  /**************************************************************************\
-   * Based on work of others;                                               *
-   * original header:                                                       *
-   *                                                                        *
-   * Copyright 2007, 2010 by Nathan C. Myers <ncm@cantrip.org>              *
-   * Redistribution and use in source and binary forms, with or without     *
-   * modification, are permitted provided that the following conditions     *
-   * are met:                                                               *
-   *                                                                        *
-   * Redistributions of source code must retain the above copyright notice, *
-   * this list of conditions and the following disclaimer.                  *
-   *                                                                        *
-   * Redistributions in binary form must reproduce the above copyright      *
-   * notice, this list of conditions and the following disclaimer in the    *
-   * documentation and/or other materials provided with the distribution.   *
-   *                                                                        *
-   * The name of the author must not be used to endorse or promote products *
-   * derived from this software without specific prior written permission.  *
-   *                                                                        *
-   * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS    *
-   * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT      *
-   * LIMITED  TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR *
-   * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT   *
-   * HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, *
-   * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT       *
-   * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,  *
-   * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY  *
-   * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT    *
-   * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE  *
-   * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.   *
-  \**************************************************************************/
-  std::pair<native_socket, native_socket> create_pipe() {
-    socklen_t addrlen = sizeof(sockaddr_in);
-    native_socket socks[2] = {invalid_native_socket, invalid_native_socket};
-    CALL_CRITICAL_CFUN(listener, cc_valid_socket, "socket",
-                       socket(AF_INET, SOCK_STREAM, IPPROTO_TCP));
-    union {
-      sockaddr_in inaddr;
-      sockaddr addr;
-    } a;
-    memset(&a, 0, sizeof(a));
-    a.inaddr.sin_family = AF_INET;
-    a.inaddr.sin_addr.s_addr = htonl(INADDR_LOOPBACK);
-    a.inaddr.sin_port = 0;
-    // makes sure all sockets are closed in case of an error
-    auto guard = detail::make_scope_guard([&] {
-      auto e = WSAGetLastError();
-      closesocket(listener);
-      closesocket(socks[0]);
-      closesocket(socks[1]);
-      WSASetLastError(e);
-    });
-    // bind listener to a local port
-    int reuse = 1;
-    CALL_CRITICAL_CFUN(tmp1, cc_zero, "setsockopt",
-                       setsockopt(listener, SOL_SOCKET, SO_REUSEADDR,
-                                  reinterpret_cast<char*>(&reuse),
-                                  static_cast<int>(sizeof(reuse))));
-    CALL_CRITICAL_CFUN(tmp2, cc_zero, "bind",
-                       bind(listener, &a.addr,
-                            static_cast<int>(sizeof(a.inaddr))));
-    // read the port in use: win32 getsockname may only set the port number
-    // (http://msdn.microsoft.com/library/ms738543.aspx):
-    memset(&a, 0, sizeof(a));
-    CALL_CRITICAL_CFUN(tmp3, cc_zero, "getsockname",
-                       getsockname(listener, &a.addr, &addrlen));
-    a.inaddr.sin_addr.s_addr = htonl(INADDR_LOOPBACK);
-    a.inaddr.sin_family = AF_INET;
-    // set listener to listen mode
-    CALL_CRITICAL_CFUN(tmp5, cc_zero, "listen", listen(listener, 1));
-    // create read-only end of the pipe
-    DWORD flags = 0;
-    CALL_CRITICAL_CFUN(read_fd, cc_valid_socket, "WSASocketW",
-                       WSASocketW(AF_INET, SOCK_STREAM, 0, nullptr, 0, flags));
-    CALL_CRITICAL_CFUN(tmp6, cc_zero, "connect",
-                       connect(read_fd, &a.addr,
-                               static_cast<int>(sizeof(a.inaddr))));
-    // get write-only end of the pipe
-    CALL_CRITICAL_CFUN(write_fd, cc_valid_socket, "accept",
-                       accept(listener, nullptr, nullptr));
-    closesocket(listener);
-    guard.disable();
-    return std::make_pair(read_fd, write_fd);
-  }
-
-#endif
 
 // -- Platform-dependent abstraction over epoll() or poll() --------------------
 
@@ -597,60 +425,8 @@ namespace network {
   }
 
 #endif // CAF_EPOLL_MULTIPLEXER
-
-// -- Helper functions for defining bitmasks of event handlers -----------------
-
-int add_flag(operation op, int bf) {
-  switch (op) {
-    case operation::read:
-      return bf | input_mask;
-    case operation::write:
-      return bf | output_mask;
-    case operation::propagate_error:
-      CAF_LOG_ERROR("unexpected operation");
-      break;
-  }
-  // weird stuff going on
-  return 0;
-}
-
-int del_flag(operation op, int bf) {
-  switch (op) {
-    case operation::read:
-      return bf & ~input_mask;
-    case operation::write:
-      return bf & ~output_mask;
-    case operation::propagate_error:
-      CAF_LOG_ERROR("unexpected operation");
-      break;
-  }
-  // weird stuff going on
-  return 0;
-}
-
-// -- Platform-independent free functions --------------------------------------
-
-expected<void> tcp_nodelay(native_socket fd, bool new_value) {
-  CAF_LOG_TRACE(CAF_ARG(fd) << CAF_ARG(new_value));
-  int flag = new_value ? 1 : 0;
-  CALL_CFUN(res, cc_zero, "setsockopt",
-            setsockopt(fd, IPPROTO_TCP, TCP_NODELAY,
-                       reinterpret_cast<setsockopt_ptr>(&flag),
-                       static_cast<socklen_t>(sizeof(flag))));
-  return unit;
-}
-
-bool is_error(ssize_t res, bool is_nonblock) {
-  if (res < 0) {
-    auto err = last_socket_error();
-    if (!is_nonblock || !would_block_or_temporarily_unavailable(err)) {
-      return true;
-    }
-    // don't report an error in case of
-    // spurious wakeup or something similar
-  }
-  return false;
-}
+  
+// -- Free functions for sending and receiving data
 
 rw_state read_some(size_t& result, native_socket fd, void* buf, size_t len) {
   CAF_LOG_TRACE(CAF_ARG(fd) << CAF_ARG(len));
@@ -677,7 +453,7 @@ rw_state write_some(size_t& result, native_socket fd, const void* buf,
   return rw_state::success;
 }
 
- bool try_accept(native_socket& result, native_socket fd) {
+bool try_accept(native_socket& result, native_socket fd) {
   CAF_LOG_TRACE(CAF_ARG(fd));
   sockaddr_storage addr;
   memset(&addr, 0, sizeof(addr));
@@ -728,20 +504,36 @@ bool write_datagram(size_t& result, native_socket fd, void* buf, size_t buf_len,
   return true;
 }
 
-// -- Policy class for TCP wrapping above free functions -----------------------
+// -- Helper functions for defining bitmasks of event handlers -----------------
 
-read_some_fun tcp_policy::read_some = network::read_some;
+int add_flag(operation op, int bf) {
+  switch (op) {
+    case operation::read:
+      return bf | input_mask;
+    case operation::write:
+      return bf | output_mask;
+    case operation::propagate_error:
+      CAF_LOG_ERROR("unexpected operation");
+      break;
+  }
+  // weird stuff going on
+  return 0;
+}
 
-write_some_fun tcp_policy::write_some = network::write_some;
-
-try_accept_fun tcp_policy::try_accept = network::try_accept;
-
-// -- Policy class for UDP wrappign above free functions -----------------------
-
-read_datagram_fun udp_policy::read_datagram = network::read_datagram;
-
-write_datagram_fun udp_policy::write_datagram = network::write_datagram;
-
+int del_flag(operation op, int bf) {
+  switch (op) {
+    case operation::read:
+      return bf & ~input_mask;
+    case operation::write:
+      return bf & ~output_mask;
+    case operation::propagate_error:
+      CAF_LOG_ERROR("unexpected operation");
+      break;
+  }
+  // weird stuff going on
+  return 0;
+}
+  
 // -- Platform-independent parts of the default_multiplexer --------------------
 
 bool default_multiplexer::try_run_once() {
@@ -990,342 +782,6 @@ int64_t default_multiplexer::next_endpoint_id() {
   return servant_ids_++;
 }
 
-event_handler::event_handler(default_multiplexer& dm, native_socket sockfd)
-    : eventbf_(0),
-      fd_(sockfd),
-      read_channel_closed_(false),
-      backend_(dm) {
-  set_fd_flags();
-}
-
-event_handler::~event_handler() {
-  if (fd_ != invalid_native_socket) {
-    CAF_LOG_DEBUG("close socket" << CAF_ARG(fd_));
-    closesocket(fd_);
-  }
-}
-
-void event_handler::close_read_channel() {
-  if (fd_ == invalid_native_socket || read_channel_closed_)
-    return;
-  ::shutdown(fd_, 0); // 0 identifies the read channel on Win & UNIX
-  read_channel_closed_ = true;
-}
-
-void event_handler::passivate() {
-  backend().del(operation::read, fd(), this);
-}
-
-void event_handler::activate() {
-  backend().add(operation::read, fd(), this);
-}
-
-void event_handler::set_fd_flags() {
-  if (fd_ == invalid_native_socket)
-    return;
-  // enable nonblocking IO, disable Nagle's algorithm, and suppress SIGPIPE
-  nonblocking(fd_, true);
-  tcp_nodelay(fd_, true);
-  allow_sigpipe(fd_, false);
-}
-
-pipe_reader::pipe_reader(default_multiplexer& dm)
-    : event_handler(dm, invalid_native_socket) {
-  // nop
-}
-
-void pipe_reader::removed_from_loop(operation) {
-  // nop
-}
-
-resumable* pipe_reader::try_read_next() {
-  intptr_t ptrval;
-  // on windows, we actually have sockets, otherwise we have file handles
-# ifdef CAF_WINDOWS
-    auto res = recv(fd(), reinterpret_cast<socket_recv_ptr>(&ptrval),
-                    sizeof(ptrval), 0);
-# else
-    auto res = read(fd(), &ptrval, sizeof(ptrval));
-# endif
-  if (res != sizeof(ptrval))
-    return nullptr;
-  return reinterpret_cast<resumable*>(ptrval);
-}
-
-void pipe_reader::handle_event(operation op) {
-  CAF_LOG_TRACE(CAF_ARG(op));
-  if (op == operation::read) {
-    auto ptr = try_read_next();
-    if (ptr != nullptr)
-      backend().resume({ptr, false});
-  }
-  // else: ignore errors
-}
-
-void pipe_reader::init(native_socket sock_fd) {
-  fd_ = sock_fd;
-}
-
-stream::stream(default_multiplexer& backend_ref, native_socket sockfd)
-    : event_handler(backend_ref, sockfd),
-      max_consecutive_reads_(
-        get_or(backend().system().config(), "middleman.max-consecutive-reads",
-               defaults::middleman::max_consecutive_reads)),
-      read_threshold_(1),
-      collected_(0),
-      ack_writes_(false),
-      writing_(false),
-      written_(0) {
-  configure_read(receive_policy::at_most(1024));
-}
-
-void stream::start(stream_manager* mgr) {
-  CAF_ASSERT(mgr != nullptr);
-  activate(mgr);
-}
-
-void stream::activate(stream_manager* mgr) {
-  if (!reader_) {
-    reader_.reset(mgr);
-    event_handler::activate();
-    prepare_next_read();
-  }
-}
-
-void stream::configure_read(receive_policy::config config) {
-  rd_flag_ = config.first;
-  max_ = config.second;
-}
-
-void stream::ack_writes(bool x) {
-  ack_writes_ = x;
-}
-
-void stream::write(const void* buf, size_t num_bytes) {
-  CAF_LOG_TRACE(CAF_ARG(num_bytes));
-  auto first = reinterpret_cast<const char*>(buf);
-  auto last  = first + num_bytes;
-  wr_offline_buf_.insert(wr_offline_buf_.end(), first, last);
-}
-
-void stream::flush(const manager_ptr& mgr) {
-  CAF_ASSERT(mgr != nullptr);
-  CAF_LOG_TRACE(CAF_ARG(wr_offline_buf_.size()));
-  if (!wr_offline_buf_.empty() && !writing_) {
-    backend().add(operation::write, fd(), this);
-    writer_ = mgr;
-    writing_ = true;
-    prepare_next_write();
-  }
-}
-
-void stream::stop_reading() {
-  CAF_LOG_TRACE("");
-  close_read_channel();
-  passivate();
-}
-
-void stream::removed_from_loop(operation op) {
-  CAF_LOG_TRACE(CAF_ARG(op));
-  switch (op) {
-    case operation::read:  reader_.reset(); break;
-    case operation::write: writer_.reset(); break;
-    case operation::propagate_error: break;
-  }
-}
-
-void stream::prepare_next_read() {
-  collected_ = 0;
-  switch (rd_flag_) {
-    case receive_policy_flag::exactly:
-      if (rd_buf_.size() != max_)
-        rd_buf_.resize(max_);
-      read_threshold_ = max_;
-      break;
-    case receive_policy_flag::at_most:
-      if (rd_buf_.size() != max_)
-        rd_buf_.resize(max_);
-      read_threshold_ = 1;
-      break;
-    case receive_policy_flag::at_least: {
-      // read up to 10% more, but at least allow 100 bytes more
-      auto max_size = max_ + std::max<size_t>(100, max_ / 10);
-      if (rd_buf_.size() != max_size)
-        rd_buf_.resize(max_size);
-      read_threshold_ = max_;
-      break;
-    }
-  }
-}
-
-void stream::prepare_next_write() {
-  CAF_LOG_TRACE(CAF_ARG(wr_buf_.size()) << CAF_ARG(wr_offline_buf_.size()));
-  written_ = 0;
-  wr_buf_.clear();
-  if (wr_offline_buf_.empty()) {
-    writing_ = false;
-    backend().del(operation::write, fd(), this);
-  } else {
-    wr_buf_.swap(wr_offline_buf_);
-  }
-}
-
-acceptor::acceptor(default_multiplexer& backend_ref, native_socket sockfd)
-    : event_handler(backend_ref, sockfd),
-      sock_(invalid_native_socket) {
-  // nop
-}
-
-void acceptor::start(acceptor_manager* mgr) {
-  CAF_LOG_TRACE(CAF_ARG2("fd", fd()));
-  CAF_ASSERT(mgr != nullptr);
-  activate(mgr);
-}
-
-void acceptor::activate(acceptor_manager* mgr) {
-  if (!mgr_) {
-    mgr_.reset(mgr);
-    event_handler::activate();
-  }
-}
-
-void acceptor::stop_reading() {
-  CAF_LOG_TRACE(CAF_ARG2("fd", fd()));
-  close_read_channel();
-  passivate();
-}
-
-void acceptor::removed_from_loop(operation op) {
-  CAF_LOG_TRACE(CAF_ARG2("fd", fd()) << CAF_ARG(op));
-  if (op == operation::read)
-    mgr_.reset();
-}
-
-datagram_handler::datagram_handler(default_multiplexer& backend_ref,
-                                   native_socket sockfd)
-    : event_handler(backend_ref, sockfd),
-      max_consecutive_reads_(
-        get_or(backend().system().config(), "middleman.max-consecutive-reads",
-               defaults::middleman::max_consecutive_reads)),
-      max_datagram_size_(receive_buffer_size),
-      rd_buf_(receive_buffer_size),
-      send_buffer_size_(0),
-      ack_writes_(false),
-      writing_(false) {
-  allow_udp_connreset(sockfd, false);
-  auto es = send_buffer_size(sockfd);
-  if (!es)
-    CAF_LOG_ERROR("cannot determine socket buffer size");
-  else
-    send_buffer_size_ = *es;
-}
-
-void datagram_handler::start(datagram_manager* mgr) {
-  CAF_LOG_TRACE(CAF_ARG2("fd", fd()));
-  CAF_ASSERT(mgr != nullptr);
-  activate(mgr);
-}
-
-void datagram_handler::activate(datagram_manager* mgr) {
-  if (!reader_) {
-    reader_.reset(mgr);
-    event_handler::activate();
-    prepare_next_read();
-  }
-}
-
-void datagram_handler::ack_writes(bool x) {
-  ack_writes_ = x;
-}
-
-
-void datagram_handler::write(datagram_handle hdl, const void* buf,
-                             size_t num_bytes) {
-  wr_offline_buf_.emplace_back();
-  wr_offline_buf_.back().first = hdl;
-  auto cbuf = reinterpret_cast<const char*>(buf);
-  wr_offline_buf_.back().second.assign(cbuf,
-                                       cbuf + static_cast<ptrdiff_t>(num_bytes));
-}
-
-void datagram_handler::flush(const manager_ptr& mgr) {
-  CAF_ASSERT(mgr != nullptr);
-  CAF_LOG_TRACE(CAF_ARG(wr_offline_buf_.size()));
-  if (!wr_offline_buf_.empty() && !writing_) {
-    backend().add(operation::write, fd(), this);
-    writer_ = mgr;
-    writing_ = true;
-    prepare_next_write();
-  }
-}
-
-
-std::unordered_map<datagram_handle, ip_endpoint>& datagram_handler::endpoints() {
-  return ep_by_hdl_;
-}
-
-const std::unordered_map<datagram_handle, ip_endpoint>&
-datagram_handler::endpoints() const {
-  return ep_by_hdl_;
-}
-
-void datagram_handler::add_endpoint(datagram_handle hdl, const ip_endpoint& ep,
-                                    const manager_ptr mgr) {
-  auto itr = hdl_by_ep_.find(ep);
-  if (itr == hdl_by_ep_.end()) {
-    hdl_by_ep_[ep] = hdl;
-    ep_by_hdl_[hdl] = ep;
-    writer_ = mgr;
-  } else if (!writer_) {
-    writer_ = mgr;
-  } else {
-    CAF_LOG_ERROR("cannot assign a second servant to the endpoint "
-                  << to_string(ep));
-    abort();
-  }
-}
-
-void datagram_handler::remove_endpoint(datagram_handle hdl) {
-  CAF_LOG_TRACE(CAF_ARG(hdl));
-  auto itr = ep_by_hdl_.find(hdl);
-  if (itr != ep_by_hdl_.end()) {
-    hdl_by_ep_.erase(itr->second);
-    ep_by_hdl_.erase(itr);
-  }
-}
-
-void datagram_handler::stop_reading() {
-  CAF_LOG_TRACE("");
-  close_read_channel();
-  passivate();
-}
-
-void datagram_handler::removed_from_loop(operation op) {
-  switch (op) {
-    case operation::read: reader_.reset(); break;
-    case operation::write: writer_.reset(); break;
-    case operation::propagate_error: break;
-  };
-}
-
-void datagram_handler::prepare_next_read() {
-  CAF_LOG_TRACE(CAF_ARG(wr_buf_.second.size())
-                << CAF_ARG(wr_offline_buf_.size()));
-  rd_buf_.resize(max_datagram_size_);
-}
-
-void datagram_handler::prepare_next_write() {
-  CAF_LOG_TRACE(CAF_ARG(wr_offline_buf_.size()));
-  wr_buf_.second.clear();
-  if (wr_offline_buf_.empty()) {
-    writing_ = false;
-    backend().del(operation::write, fd(), this);
-  } else {
-    wr_buf_.swap(wr_offline_buf_.front());
-    wr_offline_buf_.pop_front();
-  }
-}
-
 class socket_guard {
 public:
   explicit socket_guard(native_socket fd) : fd_(fd) {
@@ -1353,42 +809,6 @@ public:
 private:
   native_socket fd_;
 };
-
-auto addr_of(sockaddr_in& what) -> decltype(what.sin_addr)& {
-  return what.sin_addr;
-}
-
-auto family_of(sockaddr_in& what) -> decltype(what.sin_family)& {
-  return what.sin_family;
-}
-
-auto port_of(sockaddr_in& what) -> decltype(what.sin_port)& {
-  return what.sin_port;
-}
-
-auto addr_of(sockaddr_in6& what) -> decltype(what.sin6_addr)& {
-  return what.sin6_addr;
-}
-
-auto family_of(sockaddr_in6& what) -> decltype(what.sin6_family)& {
-  return what.sin6_family;
-}
-
-auto port_of(sockaddr_in6& what) -> decltype(what.sin6_port)& {
-  return what.sin6_port;
-}
-
-auto port_of(sockaddr& what) -> decltype(port_of(std::declval<sockaddr_in&>())) {
-  switch (what.sa_family) {
-    case AF_INET:
-      return port_of(reinterpret_cast<sockaddr_in&>(what));
-    case AF_INET6:
-      return port_of(reinterpret_cast<sockaddr_in6&>(what));
-    default:
-      break;
-  }
-  CAF_CRITICAL("invalid protocol family");
-}
 
 template <int Family>
 bool ip_connect(native_socket fd, const std::string& host, uint16_t port) {
@@ -1463,23 +883,6 @@ expected<void> set_inaddr_any(native_socket fd, sockaddr_in6& sa) {
             setsockopt(fd, IPPROTO_IPV6, IPV6_V6ONLY,
                        reinterpret_cast<setsockopt_ptr>(&off),
                        static_cast<socklen_t>(sizeof(off))));
-  return unit;
-}
-
-expected<int> send_buffer_size(native_socket fd) {
-  int size;
-  socklen_t ret_size = sizeof(size);
-  CALL_CFUN(res, cc_zero, "getsockopt",
-            getsockopt(fd, SOL_SOCKET, SO_SNDBUF,
-            reinterpret_cast<getsockopt_ptr>(&size), &ret_size));
-  return size;
-}
-
-expected<void> send_buffer_size(native_socket fd, int new_value) {
-  CALL_CFUN(res, cc_zero, "setsockopt",
-            setsockopt(fd, SOL_SOCKET, SO_SNDBUF,
-                       reinterpret_cast<setsockopt_ptr>(&new_value),
-                       static_cast<socklen_t>(sizeof(int))));
   return unit;
 }
 
@@ -1602,308 +1005,6 @@ new_local_udp_endpoint_impl(uint16_t port, const char* addr, bool reuse,
   }
   CAF_LOG_DEBUG(CAF_ARG(fd));
   return std::make_pair(fd, proto);
-}
-
-expected<std::string> local_addr_of_fd(native_socket fd) {
-  sockaddr_storage st;
-  socklen_t st_len = sizeof(st);
-  sockaddr* sa = reinterpret_cast<sockaddr*>(&st);
-  CALL_CFUN(tmp1, cc_zero, "getsockname", getsockname(fd, sa, &st_len));
-  char addr[INET6_ADDRSTRLEN] {0};
-  switch (sa->sa_family) {
-    case AF_INET:
-      return inet_ntop(AF_INET, &reinterpret_cast<sockaddr_in*>(sa)->sin_addr,
-                       addr, sizeof(addr));
-    case AF_INET6:
-      return inet_ntop(AF_INET6,
-                       &reinterpret_cast<sockaddr_in6*>(sa)->sin6_addr,
-                       addr, sizeof(addr));
-    default:
-      break;
-  }
-  return make_error(sec::invalid_protocol_family,
-                    "local_addr_of_fd", sa->sa_family);
-}
-
-expected<uint16_t> local_port_of_fd(native_socket fd) {
-  sockaddr_storage st;
-  socklen_t st_len = sizeof(st);
-  CALL_CFUN(tmp, cc_zero, "getsockname",
-            getsockname(fd, reinterpret_cast<sockaddr*>(&st), &st_len));
-  return ntohs(port_of(reinterpret_cast<sockaddr&>(st)));
-}
-
-expected<std::string> remote_addr_of_fd(native_socket fd) {
-  sockaddr_storage st;
-  socklen_t st_len = sizeof(st);
-  sockaddr* sa = reinterpret_cast<sockaddr*>(&st);
-  CALL_CFUN(tmp, cc_zero, "getpeername", getpeername(fd, sa, &st_len));
-  char addr[INET6_ADDRSTRLEN] {0};
-  switch (sa->sa_family) {
-    case AF_INET:
-      return inet_ntop(AF_INET, &reinterpret_cast<sockaddr_in*>(sa)->sin_addr,
-                       addr, sizeof(addr));
-    case AF_INET6:
-      return inet_ntop(AF_INET6,
-                       &reinterpret_cast<sockaddr_in6*>(sa)->sin6_addr,
-                       addr, sizeof(addr));
-    default:
-      break;
-  }
-  return make_error(sec::invalid_protocol_family,
-                    "remote_addr_of_fd", sa->sa_family);
-}
-
-expected<uint16_t> remote_port_of_fd(native_socket fd) {
-  sockaddr_storage st;
-  socklen_t st_len = sizeof(st);
-  CALL_CFUN(tmp, cc_zero, "getpeername",
-            getpeername(fd, reinterpret_cast<sockaddr*>(&st), &st_len));
-  return ntohs(port_of(reinterpret_cast<sockaddr&>(st)));
-}
-
-// -- default doorman and scribe implementations -------------------------------
-
-doorman_impl::doorman_impl(default_multiplexer& mx, native_socket sockfd)
-    : doorman(network::accept_hdl_from_socket(sockfd)),
-      acceptor_(mx, sockfd) {
-  // nop
-}
-
-bool doorman_impl::new_connection() {
-  CAF_LOG_TRACE("");
-  if (detached())
-    // we are already disconnected from the broker while the multiplexer
-    // did not yet remove the socket, this can happen if an I/O event causes
-    // the broker to call close_all() while the pollset contained
-    // further activities for the broker
-    return false;
-  auto& dm = acceptor_.backend();
-  auto sptr = dm.new_scribe(acceptor_.accepted_socket());
-  auto hdl = sptr->hdl();
-  parent()->add_scribe(std::move(sptr));
-  return doorman::new_connection(&dm, hdl);
-}
-
-void doorman_impl::stop_reading() {
-  CAF_LOG_TRACE("");
-  acceptor_.stop_reading();
-  detach(&acceptor_.backend(), false);
-}
-
-void doorman_impl::launch() {
-  CAF_LOG_TRACE("");
-  acceptor_.start(this);
-}
-
-std::string doorman_impl::addr() const {
-  auto x = local_addr_of_fd(acceptor_.fd());
-  if (!x)
-    return "";
-  return std::move(*x);
-}
-
-uint16_t doorman_impl::port() const {
-  auto x = local_port_of_fd(acceptor_.fd());
-  if (!x)
-    return 0;
-  return *x;
-}
-
-void doorman_impl::add_to_loop() {
-  acceptor_.activate(this);
-}
-
-void doorman_impl::remove_from_loop() {
-  acceptor_.passivate();
-}
-
-scribe_impl::scribe_impl(default_multiplexer& mx, native_socket sockfd)
-    : scribe(network::conn_hdl_from_socket(sockfd)),
-      launched_(false),
-      stream_(mx, sockfd) {
-  // nop
-}
-
-void scribe_impl::configure_read(receive_policy::config config) {
-  CAF_LOG_TRACE("");
-  stream_.configure_read(config);
-  if (!launched_)
-    launch();
-}
-
-void scribe_impl::ack_writes(bool enable) {
-  CAF_LOG_TRACE(CAF_ARG(enable));
-  stream_.ack_writes(enable);
-}
-
-std::vector<char>& scribe_impl::wr_buf() {
-  return stream_.wr_buf();
-}
-
-std::vector<char>& scribe_impl::rd_buf() {
-  return stream_.rd_buf();
-}
-
-void scribe_impl::stop_reading() {
-  CAF_LOG_TRACE("");
-  stream_.stop_reading();
-  detach(&stream_.backend(), false);
-}
-
-void scribe_impl::flush() {
-  CAF_LOG_TRACE("");
-  stream_.flush(this);
-}
-
-std::string scribe_impl::addr() const {
-  auto x = remote_addr_of_fd(stream_.fd());
-  if (!x)
-    return "";
-  return *x;
-}
-
-uint16_t scribe_impl::port() const {
-  auto x = remote_port_of_fd(stream_.fd());
-  if (!x)
-    return 0;
-  return *x;
-}
-
-void scribe_impl::launch() {
-  CAF_LOG_TRACE("");
-  CAF_ASSERT(!launched_);
-  launched_ = true;
-  stream_.start(this);
-}
-
-void scribe_impl::add_to_loop() {
-  stream_.activate(this);
-}
-
-void scribe_impl::remove_from_loop() {
-  stream_.passivate();
-}
-
-datagram_servant_impl::datagram_servant_impl(default_multiplexer& mx,
-                                             native_socket sockfd, int64_t id)
-  : datagram_servant(datagram_handle::from_int(id)),
-    launched_(false),
-    handler_(mx, sockfd) {
-  // nop
-}
-
-bool datagram_servant_impl::new_endpoint(network::receive_buffer& buf) {
-  CAF_LOG_TRACE("");
-  if (detached())
-     // we are already disconnected from the broker while the multiplexer
-     // did not yet remove the socket, this can happen if an I/O event
-     // causes the broker to call close_all() while the pollset contained
-     // further activities for the broker
-     return false;
-  // A datagram that has a source port of zero is valid and never requires a
-  // reply. In the case of CAF we can simply drop it as nothing but the
-  // handshake could be communicated which we could not reply to.
-  // Source: TCP/IP Illustrated, Chapter 10.2
-  if (network::port(handler_.sending_endpoint()) == 0)
-    return true;
-  auto& dm = handler_.backend();
-  auto hdl = datagram_handle::from_int(dm.next_endpoint_id());
-  add_endpoint(handler_.sending_endpoint(), hdl);
-  parent()->add_hdl_for_datagram_servant(this, hdl);
-  return consume(&dm, hdl, buf);
-}
-
-void datagram_servant_impl::ack_writes(bool enable) {
-  CAF_LOG_TRACE(CAF_ARG(enable));
-  handler_.ack_writes(enable);
-}
-
-std::vector<char>& datagram_servant_impl::wr_buf(datagram_handle hdl) {
-  return handler_.wr_buf(hdl);
-}
-
-void datagram_servant_impl::enqueue_datagram(datagram_handle hdl,
-                                             std::vector<char> buffer) {
-  handler_.enqueue_datagram(hdl, std::move(buffer));
-}
-
-network::receive_buffer& datagram_servant_impl::rd_buf() {
-  return handler_.rd_buf();
-}
-
-void datagram_servant_impl::stop_reading() {
-  CAF_LOG_TRACE("");
-  handler_.stop_reading();
-  detach_handles();
-  detach(&handler_.backend(), false);
-}
-
-void datagram_servant_impl::flush() {
-  CAF_LOG_TRACE("");
-  handler_.flush(this);
-}
-
-std::string datagram_servant_impl::addr() const {
-  auto x = remote_addr_of_fd(handler_.fd());
-  if (!x)
-    return "";
-  return *x;
-}
-
-uint16_t datagram_servant_impl::port(datagram_handle hdl) const {
-  auto& eps = handler_.endpoints();
-  auto itr = eps.find(hdl);
-  if (itr == eps.end())
-    return 0;
-  return network::port(itr->second);
-}
-
-uint16_t datagram_servant_impl::local_port() const {
-  auto x = local_port_of_fd(handler_.fd());
-  if (!x)
-    return 0;
-  return *x;
-}
-
-std::vector<datagram_handle> datagram_servant_impl::hdls() const {
-  std::vector<datagram_handle> result;
-  result.reserve(handler_.endpoints().size());
-  for (auto& p : handler_.endpoints())
-    result.push_back(p.first);
-  return result;
-}
-
-void datagram_servant_impl::add_endpoint(const ip_endpoint& ep,
-                                         datagram_handle hdl) {
-  handler_.add_endpoint(hdl, ep, this);
-}
-
-void datagram_servant_impl::remove_endpoint(datagram_handle hdl) {
-  handler_.remove_endpoint(hdl);
-}
-
-void datagram_servant_impl::launch() {
-  CAF_LOG_TRACE("");
-  CAF_ASSERT(!launched_);
-  launched_ = true;
-  handler_.start(this);
-}
-
-void datagram_servant_impl::add_to_loop() {
-  handler_.activate(this);
-}
-
-void datagram_servant_impl::remove_from_loop() {
-  handler_.passivate();
-}
-
-
-void datagram_servant_impl::detach_handles() {
-  for (auto& p : handler_.endpoints()) {
-    if (p.first != hdl())
-      parent()->erase(p.first);
-  }
 }
 
 } // namespace network
